+++ retail/clients/connect/client.py
@@ -62,3 +62,21 @@ class ConnectClient(RequestClient, ConnectClientInterface):
             headers=self.internal_authentication.headers,
         )
         return response.json()
+
+    def set_vtex_host_store(
+        self,
+        project_uuid: str,
+        vtex_host_store: str,
+    ) -> Dict:
+        url = (
+            f"{self.base_url}/v2/commerce/projects/"
+            f"{project_uuid}/set-vtex-host-store/"
+        )
+
+        response = self.make_request(
+            url=url,
+            method="PATCH",
+            json={"vtex_host_store": vtex_host_store},
+            headers=self.internal_authentication.headers,
+        )
+        return response.json()